services/web/requirements.txt
bs4==0.0.2
flask==3.1.0
flask_sqlalchemy==3.1.1
ijson==3.3.0
gunicorn==23.0.0
numpy==2.2.1
pandas==2.2.3
plotly==6.0.0
psycopg2-binary==2.9.10
requests==2.32.3